+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) (+ the Migration Guide),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added
+
+- The `EGaitParameterValidation2013` dataset now has the option to exclude participant "P52", which appears to have 
+  non-sensical imu data.
+
 ## [0.8.0] - 2023-30-01
 
 ### Added